Added table and importer for documents  and importer for settlements
Added table for documents and refactored the database schema to include it and join it with other tables. Added import for it and an importer for the settlements. NOTE! The settlements which have invalid mayorality code (ending with -00) are still inserted but their mayorality code and connection the the mayoralties table is null. Also fixed a bug in regions table where София(столица) didnt insert into the table due to wrong check for doubled values.