27.11.2025 UPDATE
Added pagination, replaced old statistics with the new statistics(results found/ total records in table), replaced "Няма кметство" for settlements which dont have a mayoralty with the name of the settlement. Added autosuggest dropdown with its own endpoint in the backend and word highliting i.e when you search Со and you get София in the reuslt the Со In София is highlited. Moved helper functions on frontend to a separate folder. Had to make two separate debounce functions one for searching and one for the suggestions.